// src/app/utils/types.ts
export interface Wallet {
  portfolio: { [key in keyof typeof Token]?: number | undefined };
  rates: { [key in keyof typeof Token]?: number | undefined };
}

export interface TokenValue {
  token: string | Token;
  value: number;
}

export interface ChartProps {
  pairAddress: string | undefined;
  outTokenAddress: string;
  inTokenAddress: string;
  currency: string;
}

export interface TokenCurrencyConversionRate {
  token: Token;
  currency: Currency;
  conversionRate: number;
}

export interface CurrencyOption {
  value: string;
  id: number;
  type: string;
}

export type TokenCurrencyConversionRateMap = {
  [key in keyof typeof Token]?: TokenCurrencyConversionRate | undefined;
};

export enum Sign {
  POSITIVE,
  NEGATIVE,
  NEUTRAL,
}

export enum Currency {
  EUR = 'EUR',
  USD = 'USD',
}